+++ Task_2_1_1/src/main/java/org/example/ParallelMethod.java
@@ -1,56 +1,31 @@
 package org.example;
 
+
 import java.util.ArrayList;
-import java.util.List;
 
 /**
- * Second class.
+ * Third class.
  */
 public class ParallelMethod extends ComplexNumberFinderBase {
-    private final ArrayList<Integer> rows;
-    private final int threadNum;
+    private ArrayList<Integer> row;
 
-    /**
-     * Class's constructor.
-     */
-    public ParallelMethod(ArrayList<Integer> rows, int threadNum) {
-        super("Method #2");
-        this.rows = rows;
-        this.threadNum = threadNum;
+    public ParallelMethod(ArrayList<Integer> nums) {
+        super("Method #3");
+        row = nums;
     }
 
     @Override
     public boolean hasComplexNum() {
+        boolean result = false;
         setStartTime();
-        if (rows.size() >= threadNum) {
-            List<ParallelMethodThread> numThreads = new ArrayList<>();
-            for (int i = 0; i < threadNum; i++) {
-                if (i == threadNum - 1) {
-                    numThreads.add(new ParallelMethodThread(rows.subList(rows.size() / threadNum * i
-                            , rows.size()), i));
-                } else {
-                    numThreads.add(new ParallelMethodThread(rows.subList(rows.size() / threadNum * i
-                            , rows.size() / threadNum * (i + 1) - 1), i));
-                }
-            }
+        result = row.parallelStream().anyMatch(ParallelMethod::complyCheck);
+        setEndTime();
 
-            for (ParallelMethodThread thr : numThreads) {
-                thr.getThread().start();
-            }
+        return result;
+    }
 
-
-            for (ParallelMethodThread thr : numThreads) {
-                try {
-                    thr.getThread().join();
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                }
-            }
-
-            setEndTime();
-            return Utils.findingResult;
-        } else {
-            return false;
-        }
+    private static boolean complyCheck(int num) {
+        return Utils.isComplexNum(num);
     }
 }
+
