+++ Task_2_1_1/src/main/java/org/example/ThreadMethod.java
@@ -1,31 +1,56 @@
 package org.example;
 
-
 import java.util.ArrayList;
+import java.util.List;
 
 /**
- * Third class.
+ * Second class.
  */
 public class ThreadMethod extends ComplexNumberFinderBase {
-    private ArrayList<Integer> row;
+    private final ArrayList<Integer> rows;
+    private final int threadNum;
 
-    public ThreadMethod(ArrayList<Integer> nums) {
-        super("Method #3");
-        row = nums;
+    /**
+     * Class's constructor.
+     */
+    public ThreadMethod(ArrayList<Integer> rows, int threadNum) {
+        super("Method #2");
+        this.rows = rows;
+        this.threadNum = threadNum;
     }
 
     @Override
     public boolean hasComplexNum() {
-        boolean result = false;
         setStartTime();
-        result = row.parallelStream().anyMatch(ThreadMethod::complyCheck);
-        setEndTime();
+        if (rows.size() >= threadNum) {
+            List<ThreadMethodThread> numThreads = new ArrayList<>();
+            for (int i = 0; i < threadNum; i++) {
+                if (i == threadNum - 1) {
+                    numThreads.add(new ThreadMethodThread(rows.subList(rows.size() / threadNum * i,
+                            rows.size()), i));
+                } else {
+                    numThreads.add(new ThreadMethodThread(rows.subList(rows.size() / threadNum * i,
+                            rows.size() / threadNum * (i + 1) - 1), i));
+                }
+            }
 
-        return result;
-    }
+            for (ThreadMethodThread thr : numThreads) {
+                thr.getThread().start();
+            }
 
-    private static boolean complyCheck(int num) {
-        return Utils.isComplexNum(num);
+
+            for (ThreadMethodThread thr : numThreads) {
+                try {
+                    thr.getThread().join();
+                } catch (InterruptedException e) {
+                    e.printStackTrace();
+                }
+            }
+
+            setEndTime();
+            return Utils.findingResult;
+        } else {
+            return false;
+        }
     }
 }
-
